Squash Dark UI commits from Slim
Settings: Add new configuration qualifier for UI change on the fly (3/3)

last night I got an idea after 5 weeks of breaking stuff and
doing to much complicated things. Together with mnazim we tried
today to realize this idea. So another absolut new unique SlimBean
feature is born.

The aim we wanted at the end was to give a simple and resource friendly
way to change the complete UI of a rom on the fly. So why not use
the great fallback mechanism of AOSP and change/add a new
qualifier into the AOSP core. To explain how it works.....if you want
to add a new UI just add in /res folder of any app or framework you want
folders with the qualifier "inverted". Eg. drawable-inverted-xhdpi
or values-inverted or layout-port-inverted or whatever you want. Put in
only the files you want to modify thats it.

All qualifier in AOSP are following strict rules. You can read here.
Do not forget the order of qualifier is important. The new "inverted"
qualifier we added to the UIMode qualifier. To get more information
about read here

http://developer.android.com/guide/topics/resources/providing-resources.html

So what this commit does:
- add new qualifier "inverted"
- support for API level 13 and higher
- change the custom UI on the fly
- control default UI Mode over config.xml
- working for all apps. They just need to include folders
  with the new qualifier name
- resource friendly....no resource wasting either programmticaly
  then source blow, because you just need to put in the files you
  want to change
- cause of the qualifier structure compatible with every phone,
  language or whatever qualifier you want to use on top
- 100% compatible to theme manager. If a theme misses drawables
  they are replaced with them from the current UIMode which is set.

What is the benefit why not use just a theme? We started half year
ago TRDS which was inverted gapps system base themed....means we
morphed directly the changes into the framework. This methoed was
on of the reasons why trds got so popular. Points for that are easy
first the system is themed, means each theme you activate over
theme engine works on top the themed system. So you have a base
style and can modify more if you want. The other reason is easy
a lot faster and resource friendly. Now this idea went to the
rom to provide two core system based themes and user
can do on top what they want or not without, like you have

This commit just adds the support for this

cheers

mnazim and kufikugel

Change-Id: I401dabcb3691446aa7886c8601245f2cd2db71b8

Conflicts:
	res/values/strings.xml
	res/xml/settings_headers.xml
	src/com/android/settings/Settings.java
	src/com/android/settings/slim/UserInterface.java

Settings: Kill specific apps on TRDS toggle change

to take 100% effect on the new style

Gmail, email, Playstore and contacts

Patchset:
- cleanup

Change-Id: I563307ecffe47087d19145b90b15dd58438520d8

Settings: Add to be sure also some other apps for TRDS toggle

cheers

Change-Id: I20eafdcf38bf8031d6e6d69c7403fa44bfd220b9

Conflicts:
	res/values/strings.xml
	res/xml/settings_headers.xml
	src/com/android/settings/Settings.java

Conflicts:
	res/values/cm_strings.xml
	src/com/android/settings/util/CMDProcessor.java
	src/com/android/settings/util/Helpers.java